# api/src/contracts/src/generated/types/bookingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public spa catalog, guest booking, and protected staff operations.
 * OpenAPI spec version: 0.2.0
 */

export interface BookingInput {
  serviceId: string;
  startsAt: Date;
  /**
     * @minLength 2
     * @maxLength 120
     */
  customerName: string;
  /** @maxLength 254 */
  customerEmail: string;
  /**
     * @minLength 7
     * @maxLength 40
     */
  customerPhone: string;
  /**
     * @maxLength 1000
     * @nullable
     */
  customerNote?: string | null;
  policyAccepted: true;
  idempotencyKey: string;
}