# api/src/contracts/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public spa catalog, guest booking, and protected staff operations.
 * OpenAPI spec version: 0.2.0
 */
import type { BookingStatusUpdateStatus } from './bookingStatusUpdateStatus';

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
  /**
     * @maxLength 500
     * @nullable
     */
  reason?: string | null;
}