# api/src/contracts/src/generated/types/service.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public spa catalog, guest booking, and protected staff operations.
 * OpenAPI spec version: 0.2.0
 */

export interface Service {
  id: string;
  slug: string;
  name: string;
  category: string;
  shortDescription: string;
  description: string;
  durationMinutes: number;
  /** Amount in the currency's minor unit */
  priceAmount: number;
  currency: string;
  /** @nullable */
  imageUrl: string | null;
  isFeatured: boolean;
}