# api/src/contracts/src/generated/types/spaProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public spa catalog, guest booking, and protected staff operations.
 * OpenAPI spec version: 0.2.0
 */

export interface SpaProfile {
  name: string;
  tagline: string;
  address: string;
  city: string;
  region: string;
  contactEmail: string;
  contactPhone: string;
  timezone: string;
  openingHours: string;
  cancellationPolicy: string;
}