# api/src/contracts/src/generated/types/staffAssignment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public spa catalog, guest booking, and protected staff operations.
 * OpenAPI spec version: 0.2.0
 */

export interface StaffAssignment {
  /** @nullable */
  staffId: string | null;
}