# api/src/contracts/src/generated/types/staffProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public spa catalog, guest booking, and protected staff operations.
 * OpenAPI spec version: 0.2.0
 */

export interface StaffProfile {
  id: string;
  displayName: string;
  bio: string;
  isActive: boolean;
  serviceIds: string[];
}