# web/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public spa catalog, guest booking, and protected staff operations.
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface SpaProfile {
  name: string;
  tagline: string;
  address: string;
  city: string;
  region: string;
  contactEmail: string;
  contactPhone: string;
  timezone: string;
  openingHours: string;
  cancellationPolicy: string;
}

export interface Service {
  id: string;
  slug: string;
  name: string;
  category: string;
  shortDescription: string;
  description: string;
  durationMinutes: number;
  /** Amount in the currency's minor unit */
  priceAmount: number;
  currency: string;
  /** @nullable */
  imageUrl: string | null;
  isFeatured: boolean;
}

export interface AvailabilitySlot {
  startsAt: string;
  endsAt: string;
  label: string;
}

export interface BookingInput {
  serviceId: string;
  startsAt: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  customerName: string;
  /** @maxLength 254 */
  customerEmail: string;
  /**
     * @minLength 7
     * @maxLength 40
     */
  customerPhone: string;
  /**
     * @maxLength 1000
     * @nullable
     */
  customerNote?: string | null;
  policyAccepted: true;
  idempotencyKey: string;
}

export type BookingConfirmationStatus = typeof BookingConfirmationStatus[keyof typeof BookingConfirmationStatus];


export const BookingConfirmationStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  checked_in: 'checked_in',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export interface BookingConfirmation {
  id: string;
  bookingReference: string;
  serviceName: string;
  startsAt: string;
  endsAt: string;
  durationMinutes: number;
  /** Amount in the currency's minor unit */
  priceAmount: number;
  currency: string;
  status: BookingConfirmationStatus;
}

export type ManagerBookingStatus = typeof ManagerBookingStatus[keyof typeof ManagerBookingStatus];


export const ManagerBookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  checked_in: 'checked_in',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export interface ManagerBooking {
  id: string;
  bookingReference: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  serviceName: string;
  /** @nullable */
  staffName: string | null;
  startsAt: string;
  endsAt: string;
  status: ManagerBookingStatus;
}

export interface StaffProfile {
  id: string;
  displayName: string;
  bio: string;
  isActive: boolean;
  serviceIds: string[];
}

export interface ManagerDashboard {
  todayCount: number;
  pendingCount: number;
  completedCount: number;
  nextBooking: ManagerBooking | null;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  confirmed: 'confirmed',
  checked_in: 'checked_in',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
  /**
     * @maxLength 500
     * @nullable
     */
  reason?: string | null;
}

export interface StaffAssignment {
  /** @nullable */
  staffId: string | null;
}

export interface AuditEvent {
  id: string;
  actorLabel: string;
  action: string;
  entityType: string;
  entityId: string;
  createdAt: string;
}

export type GetAvailabilityParams = {
serviceId: string;
date: string;
};

export type ListManagerBookingsParams = {
date?: string;
};